Use SDL2 integer types

+++ d3.c
@@ -3,7 +3,7 @@
 //
 
 
-short CheckIs386()
+int CheckIs386()
 {
     return 1;
 }
